chore(service): Some code refactoring
